Continued working on the project
Sorry, i completely forgot to even post this on github so a quick resume of what i've done so far.
For now, I've got a fonctional (at least it seems) system to have 3 seperated saves and a way to delete them easily as well as a fonction to either propose a new game if there is no save or continue if there is one.

For now there's some placeholders for tests (like saveium and the fact that new game simply saves the game) as well as no warning when you want to delete the save file but that's mainly for testing purpose so everything will be fixe and working when I finish up that part.

Also, I'm trying to use the least amount possible of extern sources to code this so might not be optimized and might take a bit longer to make everything since I want to do it all by myself.