functional first turret and enemy
It took a lot of time and attempts of me just trying everything possible one way and then changing direction to finally have something that works well enough (it's not set in stone right now since I still have a lot of things to do but it works like I wanted it to)

First I wanted the bullet to follow the current target of the turret but it was way too complicated and overall not necessary so I switched to simply shooting to the target but rotating the bullet with the head of the turret so it won't miss. That's why it took so many time and tries.