Added a turret shop, a start/stop button and modified some paths
It took a long time for me to figure it out but after a few days  I was finally able to make it all work, had a problem with the spawn of turret with the turret spawner but ultimately found the bug and resolved it, also added a stage number requirement to have more advanced turrets (which I did not create yet) and coins to buy the turret.
I will be moving move scripts around in the future to make it less confusing.

My next objective would be a wave system and I don't know yet how I'm gonna do that, I'll also probably add at least 1 more turret and enemy before the next update!