Waves system and second enemy.
Hi, I'm back after about a month, it's wasn't supposed to take this much between updates time but I was on vacation for about 2-3 weeks and then, when I came back, I had to clean and change somethings around in my PC, had to format everything, wait for new components which took about a week but it's all settled now, I'm making this update now so everything is saved in case something doesn't work again.

For now I simply added a system for waves, which I will expend a bit, try to balance and try to find a way to optimize a bit, I also added a second enemy which is simply a copy of the first with difference values, I will probably add a second turret and a new bullet since that's how I decided do calculate the damage.
I will also change some names for the files cause I realised that, If I want another map, I just can't use "Enemy_1_pathing" for it even tho it's called simply pathing.
Probably will also try to clean up a bit the files for more simplicity.